% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddData.R
\name{AddData}
\alias{AddData}
\title{Import Data List into PMshiny}
\usage{
AddData(dataList, overwrite = F)
}
\arguments{
\item{dataList}{EXCEL or NULL}

\item{overwrite}{default is False. if you want to overwite previous Data, set this param to True.}
}
\description{
use the AddData() fuction to import the Project Data List
}
