% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddUsers.R
\name{AddUsers}
\alias{AddUsers}
\title{Add users for PMshiny}
\usage{
AddUsers(userlist, initiate = F)
}
\arguments{
\item{userlist}{dataFrame or NULL}

\item{initiate}{default is False. if you want to add users, provide dataFrame and set this param to True.}
}
\description{
use the AddUsers() fuction to initiate the Server
}
