% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chpasswd.R
\name{chpasswd}
\alias{chpasswd}
\title{Change users password for PMshiny}
\usage{
chpasswd(username, newpasswd)
}
\arguments{
\item{username}{username for password to chage}

\item{newpasswd}{newpassword}
}
\description{
use the chpasswd() fuction to change user's password
}
