% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUser.R
\name{deleteUser}
\alias{deleteUser}
\title{Delete users for PMshiny}
\usage{
deleteUser(username)
}
\arguments{
\item{username}{user's name in database}
}
\description{
use the deleteUser() fuction to initiate the Server
}
