% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializePM.R
\name{initializePM}
\alias{initializePM}
\title{Initialize a PM server}
\usage{
initializePM(path)
}
\arguments{
\item{path}{the work direction}
}
\description{
use the initializePM() fuction to initiate the example Server
}
