% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddUsers.R
\name{AddUsers}
\alias{AddUsers}
\title{Add users for PMshiny}
\usage{
AddUsers(initiate = T, userlist)
}
\arguments{
\item{initiate}{default is T if you want to add users, provide dataFrame and set this param to True.}

\item{userlist}{a excel file,including:UserName,Password,Group,Phone and Mail}
}
\description{
use the AddUsers() fuction to add users in bulk
}
